Fix: changed the routers of some endpoint
